refactor: break store into parts
